requirements.txt
torch==2.2.2
timm
transformers==4.39.3
pillow
python-dotenv
requests
appwrite
torchvision
